first pass at making the bulk load more sharable
